Struct rustty::Style

pub struct Style(_, _);

The style of a Cell.

A Style has a Color and an Attr and represents either the foreground or background styling of a Cell.

Methods

impl Style

fn new(color: Color, attr: Attr) -> Style

Constructs a new Style with the given Color and Attr.

Examples

use rustty::{Style, Color, Attr};

let mut style = Style::new(Color::Green, Attr::BoldUnderline);
assert_eq!(style.color(), Color::Green);
assert_eq!(style.attr(), Attr::BoldUnderline);

fn with_color(c: Color) -> Style

Constructs a new Style with the given Color and the default Attr.

Examples

use rustty::{Style, Color, Attr};

let mut style = Style::with_color(Color::Cyan);
assert_eq!(style.color(), Color::Cyan);
assert_eq!(style.attr(), Attr::Default);

fn with_attr(a: Attr) -> Style

Constructs a new Style with the given Attr and the default Color.

Examples

use rustty::{Style, Color, Attr};

let mut style = Style::with_attr(Attr::UnderlineReverse);
assert_eq!(style.attr(), Attr::UnderlineReverse);
assert_eq!(style.color(), Color::Default);

fn color(&self) -> Color

Returns the Style's Color.

Examples

use rustty::{Style, Color};

let mut style = Style::with_color(Color::Yellow);
assert_eq!(style.color(), Color::Yellow);

fn set_color(&mut self, newcolor: Color) -> &mut Style

Sets the Style's Color to the given Color.

Examples

use rustty::{Style, Color};

let mut style = Style::with_color(Color::White);
assert_eq!(style.color(), Color::White);

style.set_color(Color::Black);
assert_eq!(style.color(), Color::Black);

fn attr(&self) -> Attr

Returns the Style's Attr.

Examples

use rustty::{Style, Attr};

let mut style = Style::with_attr(Attr::Reverse);
assert_eq!(style.attr(), Attr::Reverse);

fn set_attr(&mut self, newattr: Attr) -> &mut Style

Sets the given Style's Attr to the given Attr.

Examples

use rustty::{Style, Attr};

let mut style = Style::with_attr(Attr::BoldReverse);
assert_eq!(style.attr(), Attr::BoldReverse);

style.set_attr(Attr::Underline);
assert_eq!(style.attr(), Attr::Underline);