Struct rustty::Terminal

pub struct Terminal {
    // some fields omitted
}

A representation of the current terminal window.

Only one Terminal object can exist at any one time, Terminal::new() will return an Error if a Terminal object already exists. When a Terminal goes out of scope it resets the underlying terminal to its original state.

Examples

use rustty::{Terminal, Cell, Color};

// Construct a new Terminal.
let mut term = Terminal::new().unwrap();

// Terminals can be indexed to access specific cells.
// Indices are by column then row, corresponding to a cell's x and y coordinates.
term[(0, 0)] = Cell::with_char('x');
assert_eq!(term[(0, 0)].ch(), 'x');

term[(0, 1)].set_bg(Color::Red);
assert_eq!(term[(0, 1)].bg(), Color::Red);

term[(0, 2)].set_fg(Color::Blue);
assert_eq!(term[(0, 2)].fg(), Color::Blue);