Struct rustty::ui::core::frame::Frame

pub struct Frame {
    // some fields omitted
}

The Frame struct is the building block for all future widgets inside of ui. Objects of Frame abstract away the actual creation and drawing of areas of a terminal, because this process is the same for all widgets. Every widget should contain one Frame type to be used to render text to the screen

Methods

impl Frame

fn new(cols: usize, rows: usize) -> Frame

Constructs a new Frame object with a width of cols and height of rows

fn draw_into(&self, cells: &mut CellAccessor)

Draw the buffer contained inside of the base object to a valid object that implements CellAccessor.

fn resize(&mut self, new_size: Size)