Trait rustty::ui::core::painter::Painter

pub trait Painter: CellAccessor {
    fn printline_with_cell(&mut self, x: usize, y: usize, line: &str, cell: Cell) { ... }
    fn printline(&mut self, x: usize, y: usize, line: &str) { ... }
    fn halign_line(&self, line: &str, halign: HorizontalAlign, margin: usize) -> usize { ... }
    fn valign_line(&self, line: &str, valign: VerticalAlign, margin: usize) -> usize { ... }
    fn repeat_cell(&mut self, x: usize, y: usize, orientation: Orientation, count: usize, cell: Cell) { ... }
    fn draw_box(&mut self) { ... }
}

Provided Methods

fn printline_with_cell(&mut self, x: usize, y: usize, line: &str, cell: Cell)

Prints a string at the specified position.

This is a shorthand for setting each cell individually. cell's style is going to be copied to each destination cell.

Examples

use rustty::{Terminal, Cell, Color, Attr};
use rustty::ui::core::Painter;

let mut term = Terminal::new().unwrap();
let cell = Cell::with_style(Color::Default, Color::Red, Attr::Default);
term.printline_with_cell(0, 0, "foobar", cell);

fn printline(&mut self, x: usize, y: usize, line: &str)

Prints a string at the specified position.

Shorthand for printline_with_cell(x, y, line, Cell::default()).

fn halign_line(&self, line: &str, halign: HorizontalAlign, margin: usize) -> usize

Returns the proper x coord to align line in the specified halign alignment.

margin is the number of characters we want to leave near the borders.

fn valign_line(&self, line: &str, valign: VerticalAlign, margin: usize) -> usize

Returns the proper y coord to align line in the specified valign alignment.

margin is the number of characters we want to leave near the borders.

For now, the contents of line has no incidence whatsoever on the result, but when we support multi-line strings, it will, so we might as well stay consistent with halign_line().

fn repeat_cell(&mut self, x: usize, y: usize, orientation: Orientation, count: usize, cell: Cell)

fn draw_box(&mut self)