Struct rustty::ui::Canvas

pub struct Canvas {
    // some fields omitted
}

Logical clone of Frame that exposes backend functionality for users without breaking the API rules

Examples

use rustty::ui::Canvas;

let mut canvas = Canvas::new(60, 10);

let (rows, cols) = canvas.size();
// Set the entire canvas to '-' character
for i in 0..cols*rows {
    let y = i / cols;
    let x = i % cols;
    let mut cell = canvas.get_mut(x, y).unwrap();
    cell.set_ch('-');
}

Methods

impl Canvas

fn new(cols: usize, rows: usize) -> Canvas

Constructs a new Canvas object cols wide by rows high

Examples

use rustty::ui::Canvas;

let mut canvas = Canvas::new(60, 10); 

fn size(&self) -> Size

Returns the size of the canvas

Examples

use rustty::ui::Canvas;

let mut canvas = Canvas::new(60, 10);

assert_eq!(canvas.size(), (60, 10));

fn cellvec(&self) -> &Vec<Cell>

Returns a reference to the vector of cells that lie within the canvas

fn cellvec_mut(&mut self) -> &mut Vec<Cell>

Returns a mutable reference to the vector of cells that lie within the canvas

fn clear(&mut self, blank: Cell)

Clears the canvas with a blank Cell

fn pos_to_index(&self, x: usize, y: usize) -> Option<usize>

Converts a position on the screen to the relative coordinate within the Canvas cell buffer

fn get(&self, x: usize, y: usize) -> Option<&Cell>

Returns a reference to the cell at the specified position (x,*y*), in the form of an Option. If no cell exists at that position, then None

fn get_mut(&mut self, x: usize, y: usize) -> Option<&mut Cell>

Returns a mutable reference to the cell at the specified position (x, y), in the form of an Option. If no cell exists at that position, then None

fn origin(&self) -> Pos

The location of the canvas

fn set_origin(&mut self, new_origin: Pos)

Mantually sets the location of the canvas