Struct rustty::ui::HorizontalLayout

pub struct HorizontalLayout {
    // some fields omitted
}

Hold buttons and align them horizontally for drawing within a Dialog

Examples

use rustty::ui::core::{HorizontalAlign, VerticalAlign, ButtonResult, Widget, Button};
use rustty::ui::{Dialog, StdButton, HorizontalLayout};

let mut maindlg = Dialog::new(60, 10);

let b1 = StdButton::new("Quit", 'q', ButtonResult::Ok);
let b2 = StdButton::new("Foo!", 'f', ButtonResult::Custom(1));
let b3 = StdButton::new("Bar!", 'b', ButtonResult::Custom(2));

let buttons = vec![b1, b2, b3].into_iter().map(Box::new);
let buttons = buttons.map(|x| x as Box<Button>).collect();

let mut hlayout = HorizontalLayout::from_vec(buttons, 1);
hlayout.pack(&maindlg, HorizontalAlign::Middle, VerticalAlign::Bottom, (0,1));
     
maindlg.add_layout(hlayout);

Methods

impl HorizontalLayout

fn from_vec(widgets: Vec<Box<Button>>, inner_margin: usize) -> HorizontalLayout

Construct a HorizontalLayout object from a vector of boxed objects that implement Button. The current API for this function will change very soon

Examples

use rustty::ui::core::{ButtonResult, Button};
use rustty::ui::{StdButton, HorizontalLayout};

let b1 = StdButton::new("Quit", 'q', ButtonResult::Ok);
let b2 = StdButton::new("Foo!", 'f', ButtonResult::Custom(1));
let b3 = StdButton::new("Bar!", 'b', ButtonResult::Custom(2));

let v = vec![b1, b2, b3].into_iter().map(Box::new).map(|x| x as Box<Button>).collect();
let mut hlayout = HorizontalLayout::from_vec(v, 1);