Crate rustty

Rustty

Rustty is a terminal UI library that provides a simple, concise abstraction over an underlying terminal device.

Rustty is based on the concepts of cells and events. A terminal display is an array of cells, each holding a character and a set of foreground and background styles. Events are how a terminal communicates changes in its state; events are received from a terminal, processed, and pushed onto an input stream to be read and responded to.

Futher reading on the concepts behind Rustty can be found in the README

Structs

Cell
A single point on a terminal display.

Error
An error arising from terminal operations.

Terminal
A representation of the current terminal window.

Enums

Attr
The attributes of a Cell.

Color
The color of a Cell.

Event
An input event.