Struct rustty::Terminal

pub struct Terminal {
    // some fields omitted
}

A representation of the current terminal window.

Only one Terminal object can exist at any one time. When a Terminal goes out of scope it resets the underlying terminal to its original state.

Methods

impl Terminal

fn new() -> Result<Terminal, Error>

Creates a new Terminal with default settings.

fn with_char(ch: char) -> Result<Terminal, Error>

Creates a new Terminal with each cell set to the given character.

fn with_styles(fg: Style, bg: Style) -> Result<Terminal, Error>

Creates a new Terminal with each cell set to the given styles.

fn with_cell(cell: Cell) -> Result<Terminal, Error>

Creates a new Terminal with each cell set to the given cell.

fn swap_buffers(&mut self) -> Result<(), Error>

Send the current backbuffer to be displayed.

fn cols(&self) -> usize

Returns the width of the terminal.

fn rows(&self) -> usize

Returns the height of the terminal.

fn size(&self) -> (usize, usize)

Returns the size of the terminal as (cols, rows).

fn clear(&mut self) -> Result<(), Error>

Clears the internal backbuffer with whitespace.

fn clear_with_char(&mut self, ch: char) -> Result<(), Error>

Clears the internal backbuffer with the given character.

fn clear_with_styles(&mut self, fg: Style, bg: Style) -> Result<(), Error>

Clears the internal backbuffer with the given styles.

fn clear_with_cell(&mut self, cell: Cell) -> Result<(), Error>

Clears the internal backbuffer with the given cell.

fn check_resize(&self) -> bool

Checks whether the window size has changed, returning true if it has.

fn try_resize(&mut self) -> Result<Option<(usize, usize)>, Error>

fn try_resize_with_char(&mut self, ch: char) -> Result<Option<(usize, usize)>, Error>

fn try_resize_with_styles(&mut self, fg: Style, bg: Style) -> Result<Option<(usize, usize)>, Error>

fn try_resize_with_cell(&mut self, cell: Cell) -> Result<Option<(usize, usize)>, Error>

Performs a resize if the window size has changed, returning the new size. Otherwise returns none.

fn set_cursor(&mut self, c: Cursor) -> Result<(), Error>

Sets the cursor position.

fn get_event(&mut self, timeout_ms: usize) -> Result<Option<Event>, Error>