Struct rustty::Terminal

pub struct Terminal {
    // some fields omitted
}

A representation of the current terminal window.

Only one Terminal object can exist at any one time, Terminal::new() will return an Error if a Terminal object already exists. When a Terminal goes out of scope it resets the underlying terminal to its original state.

Examples

use rustty::{Terminal, Cell, Style, Color};

// Construct a new Terminal.
let mut term = Terminal::new().unwrap();

// Terminals can be indexed to access specific cells.
// Indices are by column then row, corresponding to a cell's x and y coordinates.
term[0][0] = Cell::with_char('x');
assert_eq!(term[0][0].ch(), 'x');

term[0][1].set_bg(Style::with_color(Color::Red));
assert_eq!(term[0][1].bg(), Style::with_color(Color::Red));

term[0][2].fg_mut().set_color(Color::Blue);
assert_eq!(term[0][2].fg().color(), Color::Blue);

Methods

impl Terminal

fn new() -> Result<Terminal, Error>

Constructs a new Terminal using the default Cell as a blank.

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();
assert_eq!(term[0][0].ch(), ' ');

fn with_char(ch: char) -> Result<Terminal, Error>

Constructs a new Terminal with each cell set to the given char and the default Styles.

Examples

use rustty::{Terminal, Cell};

let mut term = Terminal::with_char('x').unwrap();
assert_eq!(term[0][0].ch(), 'x');

fn with_styles(fg: Style, bg: Style) -> Result<Terminal, Error>

Constructs a new Terminal with each cell set to the given Styles and a blank char.

Examples

use rustty::{Terminal, Cell, Style, Color, Attr};

let style1 = Style::with_color(Color::Blue);
let style2 = Style::with_attr(Attr::Reverse);

let mut term = Terminal::with_styles(style1, style2).unwrap();
assert_eq!(term[0][0].fg(), Style::with_color(Color::Blue));
assert_eq!(term[0][0].bg(), Style::with_attr(Attr::Reverse));
assert_eq!(term[0][0].ch(), ' ');

fn with_cell(cell: Cell) -> Result<Terminal, Error>

Creates a new Terminal using the given cell as a blank.

Examples

use rustty::{Terminal, Cell};

let cell = Cell::with_char('x');

let mut term = Terminal::with_cell(cell).unwrap();
assert_eq!(term[0][0].ch(), 'x');

fn swap_buffers(&mut self) -> Result<(), Error>

Swaps buffers to display the current backbuffer.

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();
term.swap_buffers().unwrap();

fn cols(&self) -> usize

Returns the width of the terminal in columns.

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();
let width = term.cols();

fn rows(&self) -> usize

Returns the height of the terminal in rows.

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();
let height = term.rows();

fn size(&self) -> (usize, usize)

Returns the size of the terminal as (cols, rows).

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();
let size = term.size();
assert_eq!(size, (term.cols(), term.rows()));

fn clear(&mut self) -> Result<(), Error>

Clears the internal backbuffer with the default cell.

Examples

use rustty::{Terminal, Cell};

let mut term = Terminal::with_char('x').unwrap();
assert_eq!(term[0][0].ch(), 'x');

term.clear().unwrap();
assert_eq!(term[0][0].ch(), ' ');

fn clear_with_char(&mut self, ch: char) -> Result<(), Error>

Clears the internal backbuffer with the given char and the default Styles.

Examples

use rustty::{Terminal, Cell};

let mut term = Terminal::with_char('x').unwrap();
assert_eq!(term[0][0].ch(), 'x');

term.clear_with_char('y').unwrap();
assert_eq!(term[0][0].ch(), 'y');

fn clear_with_styles(&mut self, fg: Style, bg: Style) -> Result<(), Error>

Clears the internal backbuffer with the given Styles and a blank char.

Examples

use rustty::{Terminal, Cell, Style, Color};

let mut style1 = Style::with_color(Color::Blue);
let mut style2 = Style::with_color(Color::Red);

let mut term = Terminal::with_styles(style1, style2).unwrap();
assert_eq!(term[0][0].fg(), Style::with_color(Color::Blue));
assert_eq!(term[0][0].bg(), Style::with_color(Color::Red));

term.clear_with_styles(style2, style1).unwrap();
assert_eq!(term[0][0].fg(), Style::with_color(Color::Red));
assert_eq!(term[0][0].bg(), Style::with_color(Color::Blue));

fn clear_with_cell(&mut self, cell: Cell) -> Result<(), Error>

Clears the internal backbuffer using the given Cell as a blank.

Examples

use rustty::{Terminal, Cell};

let cell1 = Cell::with_char('x');
let cell2 = Cell::with_char('y');

let mut term = Terminal::with_cell(cell1).unwrap();
assert_eq!(term[0][0].ch(), 'x');

term.clear_with_cell(cell2).unwrap();
assert_eq!(term[0][0].ch(), 'y');

fn check_resize(&self) -> bool

Checks whether the underlying window size has changed and the buffers have not been resized yet. If this method returns true the next call to swap_buffers() or a clear() method is guaranteed to resize the buffers unless a call to a try_resize() method is made.

Examples

use rustty::{Terminal, Cell};

let mut term = Terminal::new().unwrap();

let will_resize = term.check_resize();

// If will_resize == true, swap_buffers() will resize the buffers.
term.swap_buffers().unwrap();
// So will clear().
term.clear().unwrap();

// Unless try_resize() is called.
term.try_resize().unwrap();

fn try_resize(&mut self) -> Result<Option<(usize, usize)>, Error>

Resizes the buffers if the underlying terminal window size has changed, using the default Cell as a blank.

This method will be called automatically on each call to swap_buffers() or a clear() method.

This method is guaranteed to resize the buffers if a call to check_resize() returns true and neither swap_buffers() nor a clear() method has been called since.

Returns Some((cols, rows)) if the buffers were resized and None if no resize was performed.

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();

// If new_size == Some(T) then T is the new size of the terminal.
// If new_size == None then the terminal has not resized.
let new_size = term.try_resize().unwrap();

fn try_resize_with_char(&mut self, ch: char) -> Result<Option<(usize, usize)>, Error>

Resizes the buffers if the underlying terminal window size has changed, using the given char and the default Styles as a blank.

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();

// If new_size == Some(T) then T is the new size of the terminal.
// If new_size == None then the terminal has not resized.
let new_size = term.try_resize_with_char('x').unwrap();

fn try_resize_with_styles(&mut self, fg: Style, bg: Style) -> Result<Option<(usize, usize)>, Error>

Resizes the buffers if the underlying terminal window size has changed, using the given Styles and a blank char as a blank.

Examples

use rustty::{Terminal, Style, Color};

let style = Style::with_color(Color::Red);

let mut term = Terminal::new().unwrap();

// If new_size == Some(T) then T is the new size of the terminal.
// If new_size == None then the terminal has not resized.
let new_size = term.try_resize_with_styles(style, style).unwrap();

fn try_resize_with_cell(&mut self, cell: Cell) -> Result<Option<(usize, usize)>, Error>

Resizes the buffers if the underlying terminal window size has changed, using the given Cell as a blank.

Examples

use rustty::{Terminal, Cell};

let cell = Cell::with_char('x');

let mut term = Terminal::new().unwrap();

// If new_size == Some(T) then T is the new size of the terminal.
// If new_size == None then the terminal has not resized.
let new_size = term.try_resize_with_cell(cell).unwrap();

fn set_cursor(&mut self, x: usize, y: usize) -> Result<(), Error>

Sets the cursor position to (x, y).

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();

term.set_cursor(1, 1).unwrap();

fn hide_cursor(&mut self) -> Result<(), Error>

Hides the cursor.

Examples

use rustty::Terminal;

let mut term = Terminal::new().unwrap();

term.hide_cursor().unwrap();

fn get_event(&mut self, timeout_ms: isize) -> Result<Option<Event>, Error>

Gets an event from the event stream, waiting a maximum of timeout_ms milliseconds.

Specifying a timeout_ms of -1 causes get_event() to block indefinitely, while specifying a timeout_ms of 0 causes get_event() to return immediately.

Returns Some(Event) if an event was received within the specified timeout, or None otherwise.

Examples

use rustty::{Terminal, Event};

let mut term = Terminal::new().unwrap();

let evt = term.get_event(1).unwrap();